com.unity.addressables@0.0.18-preview ## [0.0.18-preview] - 2018-04-13  - minor bug fixes  - exposed memory cache parameters to build settings, changed defaults to use LRU and timed releases to make preloading dependencies more effective ## [0.0.17-preview] - 2018-04-13  - added tests  - fixed bugs  - major API rewrite 	- all API that deals with addresses or keys have been moved to Addressables 	- LoadDependencies APIs moved to Addressables 	- Async suffix removed from all Load APIs

+++ Editor/GUI/AddressableAssetsSettingsConfigEditor.cs
@@ -83,9 +83,9 @@ namespace UnityEditor.AddressableAssets
             return false;
         }
 
-        private void DrawSpeedSlider(string text, ref uint speed)
+        private void DrawSpeedSlider(string text, ref long speed)
         {
-            speed = (uint)(Mathf.Clamp(EditorGUILayout.FloatField(text, speed / 1048576f), .1f, 1024f) * 1048576);
+            speed = (long)(Mathf.Clamp(EditorGUILayout.FloatField(text, speed / 1048576f), .1f, 1024f) * 1048576);
         }
 
         internal void OnEnable()
@@ -143,8 +143,8 @@ namespace UnityEditor.AddressableAssets
         {
             internal bool postProfilerEvents;
             internal ResourceManagerRuntimeData.EditorPlayMode editorPlayMode;
-            internal uint localLoadSpeed;
-            internal uint remoteLoadSpeed;
+            internal long localLoadSpeed;
+            internal long remoteLoadSpeed;
         }
     }
 }
